Add templatetags to render form fields
Added two templatetags for rendering the fields and widgets in a form.
For validation errors and for accesibility all the elements in a field
need to be styled correctly, not just the HTML for the renfered widget.
Normally you have to add attributes to each form widget which results
in quite a bit of boilerplate code that needs to be repeated for each
field unless custom field classes are added. Since we want to use
Block-Element-Modifier (BEM) methodology to manage the css rendering
rather than configure the widgets in code it's cleaner easier and more
maintainable to use simple tags to render template snippets. Using
separate snippets for the field and for the widget gives a greater
degress of control over the whole process.