Refactor BaseConfig to remove app settings
Using a settings file, settings.py within an app to define the default
values for the BaseConfig class was one level of indirection too many.
There several significant problem with this approach:

1. The relative path between the module containing the BaseConfig class
   and the settings module was fixed and could not be changed.

2. The code in apps.py to update the site settings by default simply
   copied over the values from the app's settings.py. As there was a
   single, shared, dict for the default values the code in apps.py needed
   to take that into account. Having custom code in apps.py that relied /
   dependend on the classes in the app was never going to work.

To solve these issues the get_initial() method was added. This returns the
initial values for the config fields which is then updated or overridden
in get_defaults(). That makes it easy to subclass.