Add template tag to load Config objects
The initial idea is that we would avoid putting logic into the templates.
As a result there was no need to add a template tag to load the various
Config classes to add conditional behaviour. However with third-party apps
we generally can't subclass the views but we can override the templates.
So loading configurations via a template tag is in.