Change "Top stories" in the site menu to "News"
There is only going to be one news section. The split between the front
page and new items seen in HackerNews and Lobste.rs is too confusing and
is really designed for sites with a large number of contributions. When
the project gets to that point then it will get revisited. Instead there
will be a main news page then an archive where stories will be grouped
by day, by week, etc.