Change settings to disable features by default
Initially the idea was to enable features by default but that ignored
the obvious problem that when a server comes online in a publicly
accessible location the attacks start immediately. Since there is always
an initial commissioning phase it's a lot safer to start with features
disabled then let the administrators switch them on when ready.